fix: update _sleep function parameter type from unsigned long to uint32_t

+++ src/wrappers/thread.cpp
@@ -1,7 +1,8 @@
 #include <chrono>
+#include <cstdint>
 #include <thread>
 
-extern "C" void _sleep(unsigned long milliseconds)
+extern "C" void _sleep(uint32_t milliseconds)
 {
     std::this_thread::sleep_for(std::chrono::milliseconds(milliseconds));
 }
